Extract feed cards into own component
