+++ src/provider/model/NotificationFeed.ts
@@ -9,6 +9,8 @@ export interface NotificationFeed {
 const PLACEHOLDER_DESCRIPTION = 'Lorem ipsum dolor sit amet, has cu dico natum. Ius veri patrioque ne. Eu mea suas appetere, quem constituto eam cu.';
 const PLACEHOLDER_LOGO = 'https://cdn.pixabay.com/photo/2016/01/26/17/15/gmail-1162901_960_720.png';
 
+// I've stored these here temporarily to keep this all together. This will be removed once
+// it is possible to get a list of feeds from the server (TODO: (SERVICE-TBD))
 export const mockFeeds: NotificationFeed[] = [
     {id: '1', subscribed: false, logo: PLACEHOLDER_LOGO, name: 'Feed 1', description: PLACEHOLDER_DESCRIPTION},
     {id: '2', subscribed: true, logo: PLACEHOLDER_LOGO, name: 'Feed 2', description: PLACEHOLDER_DESCRIPTION},
